feat: OmokController에 있던 게임 진행 로직 OmokGame 클래스로 위임
